+++ client/src/features/common/Images.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { useGetSearchRelationshipQuery } from '../../redux/api/ml_api'
 import IEntity from '../../types/data/IEntity'
-import WorkParser from '../../lib/parse/data/WorkParser'
+import EntityParser from '../../lib/parse/data/EntityParser'
 import { carriedBy } from '../../config/worksSearchTags'
 import StyledEntityPageSection from '../../styles/shared/EntityPageSection'
 import config from '../../config/config'
@@ -11,13 +11,13 @@ import IiifImageViewer from './IiifImageViewer'
 import WikiDataImageViewer from './WikiDataImageViewer'
 
 const Images: React.FC<{ entity: IEntity }> = ({ entity }) => {
-  const work = new WorkParser(entity)
-  const workCarriedByQuery = work.getHalLink(carriedBy.searchTag)
-  const skip = workCarriedByQuery === null
+  const element = new EntityParser(entity)
+  const elementCarriedByQuery = element.getHalLink(carriedBy.searchTag)
+  const skip = elementCarriedByQuery === null
   const { data, isSuccess, isLoading, isError } = useGetSearchRelationshipQuery(
     {
       // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      uri: workCarriedByQuery!,
+      uri: elementCarriedByQuery!,
     },
     {
       skip,