+++ client/src/features/landing/HeroImageSection.tsx
@@ -28,9 +28,6 @@ const HeroImageSection: React.FC<IProps> = ({ data, unit }) => {
     setImageData(landingPageImageParser.getHeroImage(unit))
   }, [data, unit])
 
-  const linkState = {
-    targetName: imageData !== null ? imageData.caption : 'Hero Image',
-  }
   return (
     <StyledHeroImageSection className="hero">
       {imageData && (
@@ -42,7 +39,6 @@ const HeroImageSection: React.FC<IProps> = ({ data, unit }) => {
           >
             <Link
               to={imageData.recordUrl}
-              state={linkState}
               onClick={() =>
                 pushClientEvent('Entity Link', 'Selected', 'Hero Image Link')
               }
@@ -62,7 +58,6 @@ const HeroImageSection: React.FC<IProps> = ({ data, unit }) => {
                       'Hero Image Link',
                     )
                   }
-                  state={linkState}
                   data-testid="hero-image-caption-link"
                 >
                   {imageData.caption.length > 30