+++ client/src/features/set/SetsPage.tsx
@@ -10,7 +10,6 @@ import StyledEntityPageSection from '../../styles/shared/EntityPageSection'
 import DataSources from '../common/DataSources'
 import EntityHeader from '../common/EntityHeader'
 import FeedbackButton from '../common/FeedbackButton'
-import WhereAtYale from '../common/WhereAtYale'
 import { ErrorFallback } from '../error/ErrorFallback'
 import UV from '../common/UV'
 import {
@@ -23,12 +22,12 @@ import { useAppDispatch } from '../../app/hooks'
 import GenericBreadcrumbHierarchy from '../common/GenericBreadcrumbHierarchy'
 import { archive } from '../../config/setsSearchTags'
 import HowDoISeeIt from '../common/HowDoISeeIt'
+import ISet from '../../types/data/ISet'
 
 import About from './About'
 import CollectionPage from './CollectionPage'
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-const SetsPage: React.FC<{ data: any }> = ({ data }) => {
+const SetsPage: React.FC<{ data: ISet }> = ({ data }) => {
   const setParser = new SetParser(data)
   const isCollectionPage =
     setParser.isClassifiedAs(config.aat.collection) ||
@@ -44,7 +43,6 @@ const SetsPage: React.FC<{ data: any }> = ({ data }) => {
     return <CollectionPage data={data} />
   }
   const isArchive = setParser.isArchive()
-  const [supertypeIcon, helperText] = setParser.getSupertypeIcon()
   const memberOf = setParser.getMemberOf()
   const objectsWithImagesHalLink = setParser.getHalLink(archive.searchTag)
   const halLinkTitle = archive.title
@@ -53,11 +51,7 @@ const SetsPage: React.FC<{ data: any }> = ({ data }) => {
     <React.Fragment>
       <div className="SetsPage">
         <ErrorBoundary FallbackComponent={ErrorFallback}>
-          <EntityHeader
-            entity={data}
-            icon={supertypeIcon}
-            entityTypeForIcon={helperText}
-          >
+          <EntityHeader entity={data}>
             {isArchive && (
               <GenericBreadcrumbHierarchy
                 key={setParser.json.id}
@@ -94,7 +88,6 @@ const SetsPage: React.FC<{ data: any }> = ({ data }) => {
             <ErrorBoundary FallbackComponent={ErrorFallback}>
               <StyledEntityPageSection className="row">
                 <HowDoISeeIt data={data} />
-                {isArchive && <WhereAtYale data={data} />}
               </StyledEntityPageSection>
               <Row>
                 <Col xs={12}>