+++ client/CHANGELOG.md
@@ -2,9 +2,29 @@
 
 All notable changes to this project will be documented in this file.
 
+## v1.37.0 - Unreleased
+
+### Added
+
+- Added on hover functionality to all icons ([#22](https://github.com/project-lux/lux-frontend/issues/22)).
+- Added support for wikidata images on object pages ([#390](https://github.com/project-lux/lux-frontend/issues/390)).
+
+### Changed
+
+- Changed Archive Sort Value AAT ([#408](https://github.com/project-lux/lux-frontend/issues/408)).
+
+### Removed
+
+- Removed Where is it at Yale? section and added its content to How do I see it? ([#249](https://github.com/project-lux/lux-frontend/issues/249)).
+
+### Fixed
+
+- Fixed rendering of entity headers ([#374](https://github.com/project-lux/lux-frontend/issues/374)).
+
 ## v1.36.1 - 2024-10-7
 
 ### Fixed
+
 - Fixed results tab redirect functionality ([#298](https://github.com/project-lux/lux-frontend/issues/298)).
 
 ## v1.36.0 - 2024-10-7